Use discovery token CA Cert hash for k8s workers
This commit introduces the use of the discovery token CA cert hash for
all the joins related to the workers nodes, which is the default
requested (and before was explicitly omitted with an unsafe option).
With this in place the branch can be merged to master since it is a
complete first step in the project of having a multi master Kubernetes
cluster deployment.