Use k8s_cidr while defining network addon
This commit introduces the k8s_cidr variable that will be used while
initializing the k8s cluster and also the network addon, by default
calico (k8s_network_addon variable).
The network addon is now a local template that is deployed on the
kubernetes directory under the home directory of the ansible user.