Use docker-ce package as default in k8s
This commit changes the default docker package to be installed when
using the kubernetes role from docker to docker-ce.
Since by default we install the docker community repositories, there's
no meaning in installing the docker package (that comes from the CentOS
repositories). Instead it makes more sense to install docker-ce.

TODO: make the docker repository a variable that can be overridden.