Introduce NotAvailable node status
This commit fixes the way node status is calculated by introducing the
NotAvailable status. First the node status is set to NotAvailable, then after
k8s_info a query for KubeletNotReady is made, and if it does not exists a query
for KubeletReady is the final action.
In this way the idempotency of the tasks should be more reliable.