Add better install explanations
This commit adds a better procedure to the README so it should be a lot easier
to prepare the environment and use it.
It also adds a note about date & time: in fact if you have a clock skew on the
machine from which you launch ansbile, there might be problems with Kube certs.